keys: Add from_nodetool_style_string factory method
Based on:
https://github.com/scylladb/scylla/commit/8daaf9833a5b3eff641f7a6cc673828d49395a5a

This patch adds a from_nodetool_style_string factory method to partition_key.
The string format is follows the nodetool format, that column in the
partition keys are split by ':'.
For example, if a partition key has two column col1 and col2, to get the
partition key that has col1 = val1 and col2 = val2:

val1:val2